[Bronze V] Title: 이상한 기호, Time: 116 ms, Memory: 15888 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,8 +7,8 @@ public class Main {
     BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
     StringTokenizer st = new StringTokenizer(br.readLine());
-    int a = Integer.parseInt(st.nextToken());
-    int b = Integer.parseInt(st.nextToken());
+    long a = Long.parseLong(st.nextToken());
+    long b = Long.parseLong(st.nextToken());
 
     bw.write((a + b) * (a - b) + "\n");
     bw.flush();
